// src/generated/models/Cr871_responsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr871_responsescr871_gateanswer = {
  144610000: 'Yes',
  144610001: 'No'
} as const;
export type Cr871_responsescr871_gateanswer = keyof typeof Cr871_responsescr871_gateanswer;
export const Cr871_responsescr871_maturitylevel = {
  144610000: 'FullyImplemented',
  144610001: 'Partial_InProgress',
  144610002: 'NotImplemented',
  144610003: 'NotApplicable',
  144610004: 'NotAnswered'
} as const;
export type Cr871_responsescr871_maturitylevel = keyof typeof Cr871_responsescr871_maturitylevel;
export const Cr871_responsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr871_responsesstatecode = keyof typeof Cr871_responsesstatecode;
export const Cr871_responsesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr871_responsesstatuscode = keyof typeof Cr871_responsesstatuscode;

export interface Cr871_responsesBase {
  "cr871_AssessmentID@odata.bind"?: string;
  cr871_compensatingcontrols?: string;
  cr871_gateanswer?: Cr871_responsescr871_gateanswer;
  cr871_iscovered?: boolean;
  cr871_isgatequestion?: boolean;
  cr871_maturitylevel?: Cr871_responsescr871_maturitylevel;
  "cr871_QuestionID@odata.bind"?: string;
  cr871_responseid: string;
  cr871_responsetext?: string;
  cr871_sectionid?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr871_responsesstatecode;
  statuscode?: Cr871_responsesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr871_responses extends Cr871_responsesBase {
  cr871_assessmentidname?: string;
  cr871_gateanswername?: string;
  cr871_iscoveredname?: string;
  cr871_isgatequestionname?: string;
  cr871_maturitylevelname?: string;
  cr871_questionidname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr871_assessmentid?: object;
  _cr871_assessmentid_value?: string;
  cr871_questionid?: object;
  _cr871_questionid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}