// src/generated/models/Cr871_questionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr871_questionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr871_questionsstatecode = keyof typeof Cr871_questionsstatecode;
export const Cr871_questionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr871_questionsstatuscode = keyof typeof Cr871_questionsstatuscode;

export interface Cr871_questionsBase {
  cr871_applicableratings?: string;
  cr871_evidencerequired?: string;
  cr871_gatequestionid?: string;
  cr871_iscoveredbycert?: boolean;
  cr871_isgatequestion?: boolean;
  cr871_questionid: string;
  cr871_questiontext?: string;
  cr871_questiontext1?: string;
  cr871_sectionid?: string;
  cr871_sortorder?: number;
  cr871_subsectionid?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr871_questionsstatecode;
  statuscode?: Cr871_questionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr871_questions extends Cr871_questionsBase {
  cr871_iscoveredbycertname?: string;
  cr871_isgatequestionname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}